Add creator_id to campaign
Summary: This adds a column to the campaign table to track which user
created the campaign, and displays that information on the admin
campaign list, as requested in #1046.

I modeled the migration off of migration #11, which added a similar
column to the message table.

I updated the two GraphQL mutations that can create campaigns,
`createCampaign` and `copyCampaign` to store the ID of the logged in
user as creator_id.

Finally, I added a `creator` field to `Campaign` in GraphQL and used
that in `CampaignList` to show which user created the campaign. Let me
know if you have feedback on how I added it to the UI, I added it to the
same line as the description since I didn't want to increase the height
of each campaign row, but I can make it it's own line if you'd prefer
that instead. I see the github issue also mentioned showing the raw ID,
and the time the campaign was created. I think I'll check with joemcl
first to make sure what I added is about what he was looking for, and
then look into adding those pieces of information.

I did notice that when adding the user_id field to the message schema,
the same commit also updated the dependencies we pass as options to
include User. I looked at createModel in rethink-knex-adapter, and it
looks like dependencies isn't used at all, so I didn't update that in
this commit, but let me know if I missed something and I should update
it!

Test Plan: Confirmed that a campaign with a null creator_id still
rendered properly, and that a campaign with a non-null creator_id showed
the correct name on the campaign list. Tested copying a campaign with no
creator_id to confirm the copied campaign had the logged-in user's id as
creator_id.