feat(tags): remove separate message editing field for tags
Using a separate editing field, similar to Opt Out, does not allow editing survey responses and
tags at the same time. This field has been removed and changes to tags are handled the same way
as survey responses are: changes are stored in state and text in the primary message text field
is updated.