fix(s3): use async method to support role-based auth
When using AWS S3 exports authenticated by an EC2 role (as is the case with Elastic Beanstalk), we
need to use the async version of getSignedUrl. The sync version does not wait for credentials to be
fetched asynchronously (as is the case when assuming a role) and the operation fails.

See aws/aws-sdk-js#2166 for more information.