fix: break props/callbacks cycles
Callbacks from the pagination components were triggering changes to the props they were passed.
This in turn kicked off another fetch by the HOC and so on. Fix this by ensuring props have
actually changed before making queries again. Make the queries directly from the Apollo client
rather than a HOC in order to have better control over when the queries are kicked off.

Ideally changes to props would cancel an existing query, but this is not really supported in
apollo-client at the moment. There are partial workarounds using AbortController and .watchQuery()
but neither work for us. See:
https://github.com/apollographql/apollo-feature-requests/issues/40#issuecomment-489487888

This also fixes a state comparison bug in AdminIncomingMessageList and fixes an Apollo complain
about a missing organization id.