feat: add promote-on-assign button (#611)
Add button for admins to both approve a request and update that user's organization membership record to be auto approve all future assignment requests. While this can be done from the people page, that requires a few extra steps when you come across a manual approval request from a texter that should be be getting auto approved.