fix: remove polling of assignment contact IDs
A change to a campaign contact's status will cause it to be excluded from the next polling of
contacts. This causes at the least the two following problems:

1. When a reply comes in the current contact index changes and the texter is booted out of the
   reply they are crafting.
2. When a message is sent and then navigated back to the texter can view the past conversation
   for a moment but is booted to a permanent loading screen when the next poll completes. The
   currentContactIndex gets set to -1 leading to no currentContact and thus the loading screen.
   Usually this could only happen while Spoke is fetching the next batch of 10 contacts.

Spoke tries to handle these cases in AssignmentTexter.componentWillUpdate but it does not really
work. A list of contacts that shifts underneath the user is also not the correct approach in my
opionion.