feat(scripteditor): add short link warning (#765)
Add a ScriptLinkWarningDialog component that is displayed when a user tries to insert a generic link or short link into an interaction script. The dialog points the user towards the relevant docs, and prompts them to confirm and save the draft script, or close the dialog and re-edit the interaction script.