build!: drop old proxy deps (#889)
Update dependencies to drop downstream dependencies on old versions of http-proxy-agent and socks-proxy-agent.

This also drops support for Mailgun as a) we are not using it and b) there are no stable versions of a node SDK.

BREAKING CHANGE: Support for Mailgun has been dropped due to no suitable Node library.

Reviewed-by: @ben-pr-p