fix: remove default onChange behavior for custom field
The field's default onChange behavior (updating the GSForm value) was being attached to the
read-only TextField. Clicking into this TextField opens the script editor pop-up but if changes are
made to the script while the TextField has focus then onChange is called with a string, rather than
an array of strings.

The script editor pop-up should steal focus as soon as it is opened, preventing/masking this bug,
but that functionality is currently broken it would appear.