test(texters): assign from all conversation states
The Texters section UI shows counts for Assigned and Unassigned but does not scope these numbers to
needsMessage conversations. The assignment query is limited to needsMessage, however, resulting in
behavior where assigning n messages to a texter in the frontend results in 0 assigned converations
because while there are n unassigned conversations none of them are needsMessage.