fix: fix interaction step management
Writing interaction steps to the cache caused the GET_EDIT_CAMPAIGN_DATA query to be re-fetched.
The results of re-fetch then overwrote the local changes for pending interaction step changes. This
re-fetch happened because in apollo-client@3 cache-and-network fetch policies re-run whenever there
is a cache update that affects the query.

For more information, see:
https://github.com/apollographql/apollo-client/issues/6760#issuecomment-668188727